endpoint arg load file renames
