+++ elrond-wasm/src/io/arg_loader_single.rs
