+++ elrond-wasm/src/io/bytes_arg_loader.rs
