+++ elrond-wasm/src/io/mod.rs
@@ -1,23 +1,23 @@
 mod arg_de_input;
+mod arg_error_handler;
 mod arg_id;
+mod arg_loader_multi;
+mod arg_loader_single;
 mod arg_nested_tuple;
-mod dyn_arg;
-mod dyn_arg_input_bytes;
-mod dyn_arg_input_endpoint;
-mod dyn_arg_input_endpoint_single;
-mod dyn_arg_input_managed_result;
+mod bytes_arg_loader;
 mod finish;
 mod macro_helpers;
+mod managed_result_arg_loader;
 mod signal_error;
 
 pub use arg_de_input::*;
+pub use arg_error_handler::*;
 pub use arg_id::ArgId;
+pub use arg_loader_multi::*;
+pub use arg_loader_single::*;
 pub use arg_nested_tuple::*;
-pub use dyn_arg::*;
-pub use dyn_arg_input_bytes::*;
-pub use dyn_arg_input_endpoint::*;
-pub use dyn_arg_input_endpoint_single::*;
-pub use dyn_arg_input_managed_result::*;
+pub use bytes_arg_loader::*;
 pub use finish::*;
 pub use macro_helpers::*;
+pub use managed_result_arg_loader::*;
 pub use signal_error::*;
